add ionic

+++ module/CodeOrders/src/CodeOrders/V1/Rest/Products/ProductsResource.php
@@ -27,7 +27,9 @@ class ProductsResource extends AbstractResourceListener
      */
     public function create($data)
     {
-        return $this->repository->insert($data);
+        $user = $this->repository->findByUsername($this->getIdentity()->getRoleId());
+        if ($user->getRole() == "salesman") return new ApiProblem(403, "The user has not access to this info");
+        return $this->repository->create($data);
     }
 
     /**
@@ -38,6 +40,8 @@ class ProductsResource extends AbstractResourceListener
      */
     public function delete($id)
     {
+        $user = $this->repository->findByUsername($this->getIdentity()->getRoleId());
+        if ($user->getRole() == "salesman") return new ApiProblem(403, "The user has not access to this info");
         return $this->repository->delete($id);
     }
 
@@ -83,7 +87,9 @@ class ProductsResource extends AbstractResourceListener
      */
     public function patch($id, $data)
     {
-        return new ApiProblem(405, 'The PATCH method has not been defined for individual resources');
+        $user = $this->repository->findByUsername($this->getIdentity()->getRoleId());
+        if ($user->getRole() == "salesman") return new ApiProblem(403, "The user has not access to this info");
+        return $this->repository->patch($id, $data);
     }
 
     /**
@@ -106,6 +112,8 @@ class ProductsResource extends AbstractResourceListener
      */
     public function update($id, $data)
     {
+        $user = $this->repository->findByUsername($this->getIdentity()->getRoleId());
+        if ($user->getRole() == "salesman") return new ApiProblem(403, "The user has not access to this info");
         return $this->repository->update($id, $data);
     }
 }
